edit: scope too large, misc edits [skip ci]
