+++ examples/PhasorDynamics/Example1/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_executable(phasordynamics_example1 example1.cpp)
 target_link_libraries(phasordynamics_example1 
-	GRIDKIT::phasor_dynamics_bus 
+	GRIDKIT::phasor_dynamics_bus
 	GRIDKIT::phasor_dynamics_bus_fault 
 	GRIDKIT::phasor_dynamics_branch
 	GRIDKIT::phasor_dynamics_genrou
