+++ examples/PhasorDynamics/Example3/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_executable(phasordynamics_example3 example3.cpp)
 target_link_libraries(phasordynamics_example3
-	GRIDKIT::phasor_dynamics_bus 
+	GRIDKIT::phasor_dynamics_bus
 	GRIDKIT::phasor_dynamics_bus_fault 
 	GRIDKIT::phasor_dynamics_branch
 	GRIDKIT::phasor_dynamics_genrou
