+++ examples/PhasorDynamics/Example3/example3.cpp
@@ -16,8 +16,8 @@
 
 // #include "example3.hpp"
 #include <Model/PhasorDynamics/Branch/Branch.hpp>
-#include <Model/PhasorDynamics/Bus/Bus.hpp>
-#include <Model/PhasorDynamics/Bus/BusInfinite/BusInfinite.hpp>
+#include <Model/PhasorDynamics/Bus/BusElectric/BusInfinite.hpp>
+#include <Model/PhasorDynamics/Bus/BusElectric/BusNetwork.hpp>
 #include <Model/PhasorDynamics/BusFault/BusFault.hpp>
 #include <Model/PhasorDynamics/Load/Load.hpp>
 #include <Model/PhasorDynamics/SynchronousMachine/GENROUwS/Genrou.hpp>
@@ -37,15 +37,15 @@ int main()
   /* Create model parts */
   BusInfinite<scalar_type, index_type> bus1(1, 0);
 
-  Bus<scalar_type, index_type> bus2(0.999733719609643, 0.0230757421433613);
-  Bus<scalar_type, index_type> bus3(0.999047460127767, 0.0436368240395443);
-  Bus<scalar_type, index_type> bus4(0.998097277498088, 0.061658775943942);
-  Bus<scalar_type, index_type> bus5(0.997021447662007, 0.0771246581966912);
-  Bus<scalar_type, index_type> bus6(0.946436885707683, 0.322888837484268);
-  Bus<scalar_type, index_type> bus7(0.943037519659334, 0.332686393642377);
-  Bus<scalar_type, index_type> bus8(0.940418229359708, 0.340019961013984);
-  Bus<scalar_type, index_type> bus9(0.938638861502395, 0.344901620288291);
-  Bus<scalar_type, index_type> bus10(0.937739191669114, 0.347340277548916);
+  BusNetwork<scalar_type, index_type> bus2(0.999733719609643, 0.0230757421433613);
+  BusNetwork<scalar_type, index_type> bus3(0.999047460127767, 0.0436368240395443);
+  BusNetwork<scalar_type, index_type> bus4(0.998097277498088, 0.061658775943942);
+  BusNetwork<scalar_type, index_type> bus5(0.997021447662007, 0.0771246581966912);
+  BusNetwork<scalar_type, index_type> bus6(0.946436885707683, 0.322888837484268);
+  BusNetwork<scalar_type, index_type> bus7(0.943037519659334, 0.332686393642377);
+  BusNetwork<scalar_type, index_type> bus8(0.940418229359708, 0.340019961013984);
+  BusNetwork<scalar_type, index_type> bus9(0.938638861502395, 0.344901620288291);
+  BusNetwork<scalar_type, index_type> bus10(0.937739191669114, 0.347340277548916);
 
   Branch<scalar_type, index_type> branch12(&bus1, &bus2, 0.001, 0.005, 0, 0);
   Branch<scalar_type, index_type> branch23(&bus2, &bus3, 0.001, 0.005, 0, 0);
