+++ examples/PhasorDynamics/Example3_GenClassical/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_executable(phasordynamics_example3_genclassical example3_genclassical.cpp)
 target_link_libraries(phasordynamics_example3_genclassical
-	GRIDKIT::phasor_dynamics_bus 
+	GRIDKIT::phasor_dynamics_bus
 	GRIDKIT::phasor_dynamics_bus_fault 
 	GRIDKIT::phasor_dynamics_branch
 	GRIDKIT::phasor_dynamics_gen_classical
