+++ src/Model/PhasorDynamics/Branch/Branch.cpp
@@ -13,7 +13,7 @@
 #include <iostream>
 
 #include <Model/PhasorDynamics/Branch/BranchData.hpp>
-#include <Model/PhasorDynamics/Bus/Bus.hpp>
+#include <Model/PhasorDynamics/Bus/BusElectric/BusNetwork.hpp>
 
 namespace GridKit
 {
