+++ src/Model/PhasorDynamics/Branch/Branch.hpp
@@ -16,7 +16,7 @@ namespace GridKit
   namespace PhasorDynamics
   {
     template <class ScalarT, typename IdxT>
-    class BusBase;
+    class BusElectric;
 
     template <typename RealT, typename IdxT>
     struct BranchData;
@@ -47,7 +47,7 @@ namespace GridKit
       using Component<ScalarT, IdxT>::f_;
 
       using real_type       = typename Component<ScalarT, IdxT>::real_type;
-      using bus_type        = BusBase<ScalarT, IdxT>;
+      using bus_type        = BusElectric<ScalarT, IdxT>;
       using model_data_type = BranchData<real_type, IdxT>;
 
     public:
