--- src/Model/PhasorDynamics/Bus/BusControl/BusSignal/CMakeLists.txt
@@ -1,12 +1,0 @@
-# [[
-#  Author(s):
-#    - Cameron Rutherford <cameron.rutherford@pnnl.gov>
-#    - Slaven Peles <peless@ornl.gov>
-#]]
-
-
-gridkit_add_library(phasor_dynamics_bus_signal
-  SOURCES
-    BusSignal.cpp
-  OUTPUT_NAME
-    gridkit_phasor_dynamics_bus_signal)
