--- src/Model/PhasorDynamics/Bus/BusControl/CMakeLists.txt
@@ -1,8 +1,0 @@
-# [[
-#  Author(s):
-#    - Luke Lowery <lukel@tamu.edu>
-#]]
-
-add_subdirectory(BusSignal)
-
-
