+++ src/Model/PhasorDynamics/Bus/BusElectric/BusInfinite.cpp
@@ -1,9 +1,10 @@
-#include "BusNetwork.hpp"
+
+#include "BusInfinite.hpp"
 
 #include <cmath>
 #include <iostream>
 
-#include <Model/PhasorDynamics/Bus/BusData.hpp>
+#include <Model/PhasorDynamics/Bus/BusElectric/BusElectricData.hpp>
 
 namespace GridKit
 {
@@ -11,68 +12,71 @@ namespace GridKit
   {
 
     /*!
-     * @brief Constructor for a phasor dynamics bus.
+     * @brief Constructor for an infinite (slack) bus.
      *
      * The model is using current balance in Cartesian coordinates.
      *
-     * @todo Arguments that should be passed to ModelEvaluatorImpl constructor:
-     * - Number of equations = 2 (size_)
-     * - Number of variables = 2 (size_)
+     * Arguments to be passed to BusBase:
+     * - Number of equations = 0 (size_)
+     * - Number of variables = 0 (size_)
      * - Number of quadratures = 0
      * - Number of optimization parameters = 0
      */
     template <class ScalarT, typename IdxT>
-    BusNetwork<ScalarT, IdxT>::BusNetwork()
-      : Vr0_(0.0), Vi0_(0.0)
+    BusInfinite<ScalarT, IdxT>::BusInfinite()
     {
-      // std::cout << "Create Bus..." << std::endl;
+      // std::cout << "Create BusInfinite..." << std::endl;
       // std::cout << "Number of equations is " << size_ << std::endl;
 
-      size_ = 2;
+      size_ = 0;
     }
 
     /*!
-     * @brief Bus constructor.
+     * @brief BusInfinite constructor.
      *
      * This constructor sets initial values for active and reactive voltage.
      *
-     * @todo Arguments that should be passed to ModelEvaluatorImpl constructor:
-     * - Number of equations = 2 (size_)
-     * - Number of variables = 2 (size_)
+     * Arguments to be passed to BusBase:
+     * - Number of equations = 0 (size_)
+     * - Number of variables = 0 (size_)
      * - Number of quadratures = 0
      * - Number of optimization parameters = 0
      */
     template <class ScalarT, typename IdxT>
-    BusNetwork<ScalarT, IdxT>::BusNetwork(ScalarT Vr, ScalarT Vi)
-      : Vr0_(Vr), Vi0_(Vi)
+    BusInfinite<ScalarT, IdxT>::BusInfinite(ScalarT Vr, ScalarT Vi)
+      : Vr_(Vr), 
+        Vi_(Vi)
     {
-      // std::cout << "Create Bus..." << std::endl;
+      // std::cout << "Create BusInfinite..." << std::endl;
       // std::cout << "Number of equations is " << size_ << std::endl;
 
-      size_ = 2;
+      size_ = 0;
     }
 
     /**
-     * @brief Construct a new Bus
+     * @brief Construct a new BusInfinite
      *
+     * Arguments to be set in BusBase:
+     * - Number of equations = 0 (size_)
+     * - Number of variables = 0 (size_)
+     * - Number of quadratures = 0
+     * - Number of optimization parameters = 0
+
      * @tparam ScalarT - type of scalar variables
      * @tparam IdxT    - type for vector/matrix indices
      * @param[in] data - structure with bus data
      */
     template <class ScalarT, typename IdxT>
-    BusNetwork<ScalarT, IdxT>::BusNetwork(const DataT& data)
-      : BusBase<ScalarT, IdxT>(data.bus_id),
-        Vr0_(data.Vr0),
-        Vi0_(data.Vi0)
+    BusInfinite<ScalarT, IdxT>::BusInfinite(const DataT& data)
+      : BusElectric<ScalarT, IdxT>(data.bus_id),
+        Vr_(data.Vr0),
+        Vi_(data.Vi0)
     {
-      // std::cout << "Create Bus..." << std::endl;
-      // std::cout << "Number of equations is " << size_ << std::endl;
-
-      size_ = 2;
+      size_ = 0;
     }
 
     template <class ScalarT, typename IdxT>
-    BusNetwork<ScalarT, IdxT>::~BusNetwork()
+    BusInfinite<ScalarT, IdxT>::~BusInfinite()
     {
       // std::cout << "Destroy PQ bus ..." << std::endl;
     }
@@ -81,32 +85,16 @@ namespace GridKit
      * @brief allocate method resizes local solution and residual vectors.
      */
     template <class ScalarT, typename IdxT>
-    int BusNetwork<ScalarT, IdxT>::allocate()
+    int BusInfinite<ScalarT, IdxT>::allocate()
     {
-      // Temporary while we use std::vector in the code
-      size_t size = static_cast<size_t>(size_);
-
-      // Resize component model data
-      f_.resize(size);
-      y_.resize(size);
-      yp_.resize(size);
-      tag_.resize(size);
-
-      fB_.resize(size);
-      yB_.resize(size);
-      ypB_.resize(size);
+      // std::cout << "Nothing to allocate for infinite bus ..." << std::endl;
 
       return 0;
     }
 
-    /*!
-     * @brief Bus variables are algebraic.
-     */
     template <class ScalarT, typename IdxT>
-    int BusNetwork<ScalarT, IdxT>::tagDifferentiable()
+    int BusInfinite<ScalarT, IdxT>::tagDifferentiable()
     {
-      tag_[0] = false;
-      tag_[1] = false;
       return 0;
     }
 
@@ -114,30 +102,31 @@ namespace GridKit
      * @brief initialize method sets bus variables to stored initial values.
      */
     template <class ScalarT, typename IdxT>
-    int BusNetwork<ScalarT, IdxT>::initialize()
+    int BusInfinite<ScalarT, IdxT>::initialize()
     {
-      // std::cout << "Initialize Bus..." << std::endl;
-      y_[0]  = Vr0_;
-      y_[1]  = Vi0_;
-      yp_[0] = 0.0;
-      yp_[1] = 0.0;
+      // std::cout << "Initialize BusInfinite..." << std::endl;
 
       return 0;
     }
 
     /*!
-     * @brief PQ bus does not compute residuals, so here we just reset residual values.
+     * @brief Reset slack currents to zero.
+     *
+     * Infinite bus does not compute residuals, so here we just reset
+     * current values to zero. Components connected to the infinite bus
+     * will add their currents to Ir_ and Ii_. The resultant will be slack
+     * current that the infinite bus has to pick up.
      *
      * @warning This implementation assumes bus residuals are always evaluated
      * _before_ component model residuals.
      *
      */
     template <class ScalarT, typename IdxT>
-    int BusNetwork<ScalarT, IdxT>::evaluateResidual()
+    int BusInfinite<ScalarT, IdxT>::evaluateResidual()
     {
       // std::cout << "Evaluating residual of a PQ bus ...\n";
-      f_[0] = 0.0;
-      f_[1] = 0.0;
+      Ir_ = 0.0;
+      Ii_ = 0.0;
       return 0;
     }
 
@@ -149,7 +138,7 @@ namespace GridKit
      * @return int - error code
      */
     template <class ScalarT, typename IdxT>
-    int BusNetwork<ScalarT, IdxT>::evaluateJacobian()
+    int BusInfinite<ScalarT, IdxT>::evaluateJacobian()
     {
       return 0;
     }
@@ -158,30 +147,23 @@ namespace GridKit
      * @brief initialize method sets bus variables to stored initial values.
      */
     template <class ScalarT, typename IdxT>
-    int BusNetwork<ScalarT, IdxT>::initializeAdjoint()
+    int BusInfinite<ScalarT, IdxT>::initializeAdjoint()
     {
-      // std::cout << "Initialize Bus..." << std::endl;
-      yB_[0]  = 0.0;
-      yB_[1]  = 0.0;
-      ypB_[0] = 0.0;
-      ypB_[1] = 0.0;
+      // std::cout << "Initialize BusInfinite..." << std::endl;
 
       return 0;
     }
 
     /**
-     * @brief Bus only initializes adjoint residual elements to zero.
+     * @brief BusInfinite only initializes adjoint residual elements to zero.
      *
      * @tparam ScalarT - data type for the integrand
      * @tparam IdxT    - data type for matrix/vector indices
      * @return int - error code
      */
     template <class ScalarT, typename IdxT>
-    int BusNetwork<ScalarT, IdxT>::evaluateAdjointResidual()
+    int BusInfinite<ScalarT, IdxT>::evaluateAdjointResidual()
     {
-      fB_[0] = 0.0;
-      fB_[1] = 0.0;
-
       return 0;
     }
 
@@ -193,7 +175,7 @@ namespace GridKit
      * @return int - error code
      */
     template <class ScalarT, typename IdxT>
-    int BusNetwork<ScalarT, IdxT>::evaluateIntegrand()
+    int BusInfinite<ScalarT, IdxT>::evaluateIntegrand()
     {
       return 0;
     }
@@ -206,10 +188,16 @@ namespace GridKit
      * @return int - error code
      */
     template <class ScalarT, typename IdxT>
-    int BusNetwork<ScalarT, IdxT>::evaluateAdjointIntegrand()
+    int BusInfinite<ScalarT, IdxT>::evaluateAdjointIntegrand()
     {
       return 0;
     }
+
+    // Available template instantiations
+    template class BusInfinite<double, long int>;
+    template class BusInfinite<double, size_t>;
+    template class BusInfinite<DependencyTracking::Variable, long int>;
+    template class BusInfinite<DependencyTracking::Variable, size_t>;
 
   } // namespace PhasorDynamics
 } // namespace GridKit
