--- src/Model/PhasorDynamics/Bus/BusElectric/BusInfinite/README.md
@@ -1,3 +1,0 @@
-# Infinite Bus
-
-An infinite bus is an electric bus that is an ideal voltage source.
