+++ src/Model/PhasorDynamics/Bus/BusElectric/BusNetwork.cpp
@@ -1,38 +1,56 @@
-
-/**
- * @file BusSignal.cpp
- * @author Luke Lowery (lukel@tamu.edu)
- * @brief Definition of BusSignal class.
- *
- */
-
-#include "BusSignal.hpp"
+#include "BusNetwork.hpp"
 
 #include <cmath>
 #include <iostream>
 
-#include <Model/PhasorDynamics/Bus/BusData.hpp>
-
+#include <Model/PhasorDynamics/Bus/BusElectric/BusElectricData.hpp>
 namespace GridKit
 {
   namespace PhasorDynamics
   {
 
-        /**
-     * @brief Construct a new Bus
+    /*!
+     * @brief Constructor for a phasor dynamics bus.
      *
-     * @tparam ScalarT - type of scalar variables
-     * @tparam IdxT    - type for vector/matrix indices
-     * @param[in] data - structure with bus data
+     * The model is using current balance in Cartesian coordinates.
+     *
+     * @todo Arguments that should be passed to ModelEvaluatorImpl constructor:
+     * - Number of equations = 2 (size_)
+     * - Number of variables = 2 (size_)
+     * - Number of quadratures = 0
+     * - Number of optimization parameters = 0
      */
     template <class ScalarT, typename IdxT>
-    BusSignal<ScalarT, IdxT>::BusSignal()
-      : BusSignalBase<ScalarT, IdxT>()
+    BusNetwork<ScalarT, IdxT>::BusNetwork()
+      : Vr0_(0.0), 
+        Vi0_(0.0)
     {
       // std::cout << "Create Bus..." << std::endl;
       // std::cout << "Number of equations is " << size_ << std::endl;
 
-      size_ = 0;
+      size_ = 2;
+    }
+
+    /*!
+     * @brief Bus constructor.
+     *
+     * This constructor sets initial values for active and reactive voltage.
+     *
+     * @todo Arguments that should be passed to ModelEvaluatorImpl constructor:
+     * - Number of equations = 2 (size_)
+     * - Number of variables = 2 (size_)
+     * - Number of quadratures = 0
+     * - Number of optimization parameters = 0
+     */
+    template <class ScalarT, typename IdxT>
+    BusNetwork<ScalarT, IdxT>::BusNetwork(ScalarT Vr, ScalarT Vi)
+      : Vr0_(Vr), 
+        Vi0_(Vi)
+    {
+      // std::cout << "Create Bus..." << std::endl;
+      // std::cout << "Number of equations is " << size_ << std::endl;
+
+      size_ = 2;
     }
 
     /**
@@ -43,17 +61,19 @@ namespace GridKit
      * @param[in] data - structure with bus data
      */
     template <class ScalarT, typename IdxT>
-    BusSignal<ScalarT, IdxT>::BusSignal(const DataT& data)
-      : BusSignalBase<ScalarT, IdxT>(data.bus_id)
+    BusNetwork<ScalarT, IdxT>::BusNetwork(const DataT& data)
+      : BusElectric<ScalarT, IdxT>(data.bus_id),
+        Vr0_(data.Vr0), 
+        Vi0_(data.Vi0)
     {
       // std::cout << "Create Bus..." << std::endl;
       // std::cout << "Number of equations is " << size_ << std::endl;
 
-      size_ = 0;
+      size_ = 2;
     }
 
     template <class ScalarT, typename IdxT>
-    BusSignal<ScalarT, IdxT>::~BusSignal()
+    BusNetwork<ScalarT, IdxT>::~BusNetwork()
     {
       // std::cout << "Destroy PQ bus ..." << std::endl;
     }
@@ -62,7 +82,7 @@ namespace GridKit
      * @brief allocate method resizes local solution and residual vectors.
      */
     template <class ScalarT, typename IdxT>
-    int BusSignal<ScalarT, IdxT>::allocate()
+    int BusNetwork<ScalarT, IdxT>::allocate()
     {
       // Temporary while we use std::vector in the code
       size_t size = static_cast<size_t>(size_);
@@ -72,9 +92,11 @@ namespace GridKit
       y_.resize(size);
       yp_.resize(size);
       tag_.resize(size);
+
       fB_.resize(size);
       yB_.resize(size);
       ypB_.resize(size);
+
       return 0;
     }
 
@@ -82,8 +104,10 @@ namespace GridKit
      * @brief Bus variables are algebraic.
      */
     template <class ScalarT, typename IdxT>
-    int BusSignal<ScalarT, IdxT>::tagDifferentiable()
+    int BusNetwork<ScalarT, IdxT>::tagDifferentiable()
     {
+      tag_[0] = false;
+      tag_[1] = false;
       return 0;
     }
 
@@ -91,9 +115,14 @@ namespace GridKit
      * @brief initialize method sets bus variables to stored initial values.
      */
     template <class ScalarT, typename IdxT>
-    int BusSignal<ScalarT, IdxT>::initialize()
+    int BusNetwork<ScalarT, IdxT>::initialize()
     {
       // std::cout << "Initialize Bus..." << std::endl;
+      y_[0]  = Vr0_;
+      y_[1]  = Vi0_;
+      yp_[0] = 0.0;
+      yp_[1] = 0.0;
+
       return 0;
     }
 
@@ -105,9 +134,11 @@ namespace GridKit
      *
      */
     template <class ScalarT, typename IdxT>
-    int BusSignal<ScalarT, IdxT>::evaluateResidual()
+    int BusNetwork<ScalarT, IdxT>::evaluateResidual()
     {
       // std::cout << "Evaluating residual of a PQ bus ...\n";
+      f_[0] = 0.0;
+      f_[1] = 0.0;
       return 0;
     }
 
@@ -119,7 +150,7 @@ namespace GridKit
      * @return int - error code
      */
     template <class ScalarT, typename IdxT>
-    int BusSignal<ScalarT, IdxT>::evaluateJacobian()
+    int BusNetwork<ScalarT, IdxT>::evaluateJacobian()
     {
       return 0;
     }
@@ -128,9 +159,14 @@ namespace GridKit
      * @brief initialize method sets bus variables to stored initial values.
      */
     template <class ScalarT, typename IdxT>
-    int BusSignal<ScalarT, IdxT>::initializeAdjoint()
+    int BusNetwork<ScalarT, IdxT>::initializeAdjoint()
     {
       // std::cout << "Initialize Bus..." << std::endl;
+      yB_[0]  = 0.0;
+      yB_[1]  = 0.0;
+      ypB_[0] = 0.0;
+      ypB_[1] = 0.0;
+
       return 0;
     }
 
@@ -142,8 +178,11 @@ namespace GridKit
      * @return int - error code
      */
     template <class ScalarT, typename IdxT>
-    int BusSignal<ScalarT, IdxT>::evaluateAdjointResidual()
+    int BusNetwork<ScalarT, IdxT>::evaluateAdjointResidual()
     {
+      fB_[0] = 0.0;
+      fB_[1] = 0.0;
+
       return 0;
     }
 
@@ -155,7 +194,7 @@ namespace GridKit
      * @return int - error code
      */
     template <class ScalarT, typename IdxT>
-    int BusSignal<ScalarT, IdxT>::evaluateIntegrand()
+    int BusNetwork<ScalarT, IdxT>::evaluateIntegrand()
     {
       return 0;
     }
@@ -168,10 +207,16 @@ namespace GridKit
      * @return int - error code
      */
     template <class ScalarT, typename IdxT>
-    int BusSignal<ScalarT, IdxT>::evaluateAdjointIntegrand()
+    int BusNetwork<ScalarT, IdxT>::evaluateAdjointIntegrand()
     {
       return 0;
     }
+
+    // Available template instantiations
+    template class BusNetwork<double, long int>;
+    template class BusNetwork<double, size_t>;
+    template class BusNetwork<DependencyTracking::Variable, long int>;
+    template class BusNetwork<DependencyTracking::Variable, size_t>;
 
   } // namespace PhasorDynamics
 } // namespace GridKit
