--- src/Model/PhasorDynamics/Bus/BusElectric/BusNetwork/CMakeLists.txt
@@ -1,11 +1,0 @@
-# [[
-#  Author(s):
-#    - Luke Lowery <lukel@tamu.edu>
-#]]
-
-
-gridkit_add_library(phasor_dynamics_bus_network
-  SOURCES
-    BusNetwork.cpp
-  OUTPUT_NAME
-    gridkit_phasor_dynamics_bus_network)
