--- src/Model/PhasorDynamics/Bus/BusElectric/BusNetwork/README.md
@@ -1,3 +1,0 @@
-# Network Bus
-
-A network bus is a traditional bus in an electrical network subject to KVL and KCL laws.
