+++ src/Model/PhasorDynamics/Bus/CMakeLists.txt
@@ -3,7 +3,11 @@
 #    - Luke Lowery <lukel@tamu.edu>
 #]]
 
-add_subdirectory(BusElectric)
-add_subdirectory(BusControl)
 
-
+gridkit_add_library(phasor_dynamics_bus
+  SOURCES
+    BusElectric/BusNetwork.cpp
+    BusElectric/BusInfinite.cpp
+    BusControl/BusSignal.cpp
+  OUTPUT_NAME
+    gridkit_phasor_dynamics_bus)
