+++ src/Model/PhasorDynamics/Load/Load.hpp
@@ -10,7 +10,7 @@ namespace GridKit
   namespace PhasorDynamics
   {
     template <class ScalarT, typename IdxT>
-    class BusBase;
+    class BusElectric;
 
     template <typename RealT, typename IdxT>
     struct LoadData;
@@ -40,7 +40,7 @@ namespace GridKit
       using Component<ScalarT, IdxT>::component_id_;
 
       using real_type       = typename Component<ScalarT, IdxT>::real_type;
-      using bus_type        = BusBase<ScalarT, IdxT>;
+      using bus_type        = BusElectric<ScalarT, IdxT>;
       using model_data_type = LoadData<real_type, IdxT>;
 
     public:
