+++ src/Model/PhasorDynamics/SynchronousMachine/GENROUwS/Genrou.hpp
@@ -16,7 +16,7 @@ namespace GridKit
   namespace PhasorDynamics
   {
     template <class ScalarT, typename IdxT>
-    class BusBase;
+    class BusElectric;
 
     template <typename RealT, typename IdxT>
     struct GenrouData;
@@ -47,7 +47,7 @@ namespace GridKit
       using Component<ScalarT, IdxT>::ypB_;
 
       using real_type       = typename Component<ScalarT, IdxT>::real_type;
-      using bus_type        = BusBase<ScalarT, IdxT>;
+      using bus_type        = BusElectric<ScalarT, IdxT>;
       using model_data_type = GenrouData<real_type, IdxT>;
 
     public:
