+++ tests/UnitTests/PhasorDynamics/BranchTests.hpp
@@ -3,8 +3,8 @@
 
 #include <AutomaticDifferentiation/DependencyTracking/Variable.hpp>
 #include <Model/PhasorDynamics/Branch/Branch.hpp>
-#include <Model/PhasorDynamics/Bus/Bus.hpp>
-#include <Model/PhasorDynamics/Bus/BusInfinite/BusInfinite.hpp>
+#include <Model/PhasorDynamics/Bus/BusElectric/BusNetwork.hpp>
+#include <Model/PhasorDynamics/Bus/BusElectric/BusInfinite.hpp>
 #include <Utilities/TestHelpers.hpp>
 #include <Utilities/Testing.hpp>
 
@@ -27,8 +27,8 @@ namespace GridKit
       {
         TestStatus success = true;
 
-        auto* bus1 = new PhasorDynamics::Bus<ScalarT, IdxT>(1.0, 0.0);
-        auto* bus2 = new PhasorDynamics::Bus<ScalarT, IdxT>(1.0, 0.1);
+        auto* bus1 = new PhasorDynamics::BusNetwork<ScalarT, IdxT>(1.0, 0.0);
+        auto* bus2 = new PhasorDynamics::BusNetwork<ScalarT, IdxT>(1.0, 0.1);
 
         PhasorDynamics::Component<ScalarT, IdxT>* branch =
             new PhasorDynamics::Branch<ScalarT, IdxT>(bus1, bus2);
